+++ djangocms_content_expiry/constants.py
@@ -2,3 +2,5 @@ from django.utils.translation import ugettext_lazy as _
 
 
 CONTENT_EXPIRY_EXPIRE_FIELD_LABEL = _("expiry date")
+
+CONTENT_EXPIRY_CHANGELIST_PAGECONTENT_EXCLUSION_CACHE_KEY = "djangocms_content_expiry_changelist_pagecontent_exclusion"